requirements.txt
fastapi==0.106.0
uvicorn == 0.23.1
sqlalchemy == 2.0.23
pydantic-settings==2.0.3
asyncpg==0.28.0
orjson==3.9.7
python-jose==3.3.0
pytest-asyncio==0.20.3
httpx==0.25.2
prometheus_client==0.19.0